# online-exam-platform/services/exam-service/src/main/java/com/examplatform/examservice/dto/AnswerResponse.java
package com.examplatform.examservice.dto;

import java.time.Instant;

public record AnswerResponse(
        String questionId,
        String answer,
        boolean autoSaved,
        int version,
        Instant savedAt) {
}