# online-exam-platform/services/exam-service/src/main/java/com/examplatform/examservice/dto/AnswerSubmission.java
package com.examplatform.examservice.dto;

import jakarta.validation.constraints.NotBlank;

public record AnswerSubmission(
        @NotBlank String questionId,
        @NotBlank String answer,
        Integer version) {
}