# online-exam-platform/services/exam-service/src/main/java/com/examplatform/examservice/dto/SessionResponse.java
package com.examplatform.examservice.dto;

import java.time.Instant;

public record SessionResponse(
        String studentId,
        String examId,
        String status,
        Instant startTime,
        Instant endTime,
        long timeRemaining,
        int answeredCount) {
}